+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changelog
 * `ContentBuilder` struct removed.
 * Used language assignment must be assigned again in one template.
 * Template, language and other keys are not generic anymore. All keys are `String` now.
+* Attempt to define invalid template group now return `Result::Err` and `Terarium::check_group_config_validity` 
+is removed from interface.
+* `TerariumBuilder::add_template()` now return `Result<Self, TerariumBuilderError>` to get symmetry with
+the `TerariumBuilder::add_group()` method.
 
 ## 0.1